+++ src/jvmMain/kotlin/space/kscience/gdml/streamIO.kt
@@ -3,6 +3,7 @@ package space.kscience.gdml
 import nl.adaptivity.xmlutil.StAXReader
 import nl.adaptivity.xmlutil.StAXWriter
 import nl.adaptivity.xmlutil.XmlDeclMode
+import java.io.File
 import java.io.InputStream
 import java.io.OutputStream
 import java.nio.file.Files
@@ -16,29 +17,35 @@ public fun Gdml.Companion.decodeFromStream(stream: InputStream, usePreprocessor:
     val xmlReader = StAXReader(stream, "UTF-8")
     return if (usePreprocessor) {
         val preprocessor = GdmlPreprocessor(xmlReader) { parseAndEvaluate(it) }
-        format.decodeFromReader(serializer(), preprocessor)
+        xmlFormat.decodeFromReader(serializer(), preprocessor)
     } else {
-        format.decodeFromReader(serializer(), xmlReader)
+        xmlFormat.decodeFromReader(serializer(), xmlReader)
     }
 }
 
 /**
  * Parse the file at [Path] optionally using variable pre-processor with [usePreprocessor] flag
+ *
+ * **NOTE**: Preprocessor options could be expanded in future
  */
-public fun Gdml.Companion.decodeFromFile(path: Path, usePreprocessor: Boolean = false): Gdml {
-    return Files.newInputStream(path, StandardOpenOption.READ).use {
-        decodeFromStream(it, usePreprocessor)
-    }
+public fun Gdml.Companion.decodeFromFile(
+    path: Path,
+    usePreprocessor: Boolean = false,
+): Gdml = Files.newInputStream(path, StandardOpenOption.READ).use {
+    decodeFromStream(it, usePreprocessor)
 }
+
+public fun Gdml.Companion.decodeFromFile(file: File, usePreprocessor: Boolean = false): Gdml =
+    decodeFromFile(file.toPath(), usePreprocessor = false)
 
 public fun Gdml.encodeToStream(stream: OutputStream) {
     val xmlWriter = StAXWriter(stream, "UTF-8", false, XmlDeclMode.Auto)
-    Gdml.format.encodeToWriter(xmlWriter, Gdml.serializer(), this)
+    Gdml.xmlFormat.encodeToWriter(xmlWriter, Gdml.serializer(), this)
+    stream.flush()
 }
 
 public fun Gdml.encodeToFile(path: Path) {
     Files.newOutputStream(path, StandardOpenOption.CREATE_NEW, StandardOpenOption.WRITE).use {
         encodeToStream(it)
-        it.flush()
     }
 }